fix: actually call runNow
Previously we were simply `bind`ing the `payload` for the call to
`processShard`, but we weren't actually calling it. This fixes that.

To ensure this doesn't happen again I'm adding a more substantial test
that will test the whole code base and would have caught this issue.

To write the test I've upgraded various dependencies to make the tests
run a little faster and to make the test output a little more readable,
as well as give us better fake timer support.